move module config to module file
